% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{lynch_palette}
\alias{lynch_palette}
\title{A David Lynch palette generator}
\usage{
lynch_palette(name, n)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{blue_velvet}, \code{dune},  \code{lost_highway},  \code{mullholland_drive},  \code{twin_peaks},  \code{wild_at_heart}.}

\item{n}{Number of colors desired. All palettes have 5 colors. Most color
schemes are derived from \href{https://www.pinterest.se/pin/151152131226891232/}{this Pin on Pinterest}.
If omitted, uses all colours.}
}
\value{
A vector of colours.
}
\description{
Color palettes from David Lynch films.
}
\examples{
lynch_palette("twin_peaks")
lynch_palette("blue_velvet", 3)
}
\keyword{colors}
